api/serializers.py
from django.contrib.auth import get_user_model
from rest_framework import serializers


User = get_user_model()


class UserSignUpSerializer(serializers.Serializer):
    email = serializers.EmailField()
    password = serializers.CharField(write_only=True, min_length=6)

    def validate_email(self, value: str) -> str:
        """Check if email is already registered."""
        if User.objects.filter(email=value).exists() or User.objects.filter(username=value).exists():
            raise serializers.ValidationError("A user with this email already exists.")
        return value

