requirements.txt
asgiref>=3.11.0
Django>=5.2.0,<6.0
djangorestframework>=3.15.0
drf-spectacular>=0.28.0
psycopg2-binary>=2.9.0
python-dotenv>=1.2.0
sqlparse>=0.5.0
tzdata>=2025.3

